Rename to ArborX_GeometryAlgorithms.hpp
