+++ src/cluster/detail/ArborX_MinimumSpanningTree.hpp
@@ -12,7 +12,7 @@
 #ifndef ARBORX_DETAILS_MINIMUM_SPANNING_TREE_HPP
 #define ARBORX_DETAILS_MINIMUM_SPANNING_TREE_HPP
 
-#include <detail/ArborX_Algorithms.hpp>
+#include <detail/ArborX_GeometryAlgorithms.hpp>
 #include <detail/ArborX_HappyTreeFriends.hpp>
 #include <detail/ArborX_WeightedEdge.hpp>
 #include <kokkos_ext/ArborX_KokkosExtArithmeticTraits.hpp>
