+++ src/cluster/detail/ArborX_MutualReachabilityDistance.hpp
@@ -13,7 +13,7 @@
 #define ARBORX_DETAILS_MUTUAL_REACHABILITY_DISTANCE_HPP
 
 #include <detail/ArborX_AccessTraits.hpp>
-#include <detail/ArborX_Algorithms.hpp>
+#include <detail/ArborX_GeometryAlgorithms.hpp>
 #include <detail/ArborX_HappyTreeFriends.hpp>
 #include <detail/ArborX_Predicates.hpp>
