+++ src/interpolation/detail/ArborX_InterpDetailsCompactRadialBasisFunction.hpp
@@ -13,7 +13,7 @@
 #define ARBORX_INTERP_DETAILS_COMPACT_RADIAL_BASIS_FUNCTION_HPP
 
 #include <ArborX_GeometryTraits.hpp>
-#include <detail/ArborX_Algorithms.hpp>
+#include <detail/ArborX_GeometryAlgorithms.hpp>
 
 #include <Kokkos_Core.hpp>
