+++ src/spatial/detail/ArborX_TreeTraversal.hpp
@@ -11,7 +11,7 @@
 #ifndef ARBORX_DETAILS_TREE_TRAVERSAL_HPP
 #define ARBORX_DETAILS_TREE_TRAVERSAL_HPP
 
-#include <detail/ArborX_Algorithms.hpp>
+#include <detail/ArborX_GeometryAlgorithms.hpp>
 #include <detail/ArborX_HappyTreeFriends.hpp>
 #include <detail/ArborX_NearestBufferProvider.hpp>
 #include <detail/ArborX_Node.hpp> // ROPE_SENTINEL
