+++ test/ArborXTest_LegacyTree.hpp
@@ -13,8 +13,8 @@
 #define ARBORX_TEST_LEGACY_TREE_HPP
 
 #include <detail/ArborX_AccessTraits.hpp>
-#include <detail/ArborX_Algorithms.hpp>
 #include <detail/ArborX_Callbacks.hpp>
+#include <detail/ArborX_GeometryAlgorithms.hpp>
 #include <detail/ArborX_PairValueIndex.hpp>
 
 #include <Kokkos_Macros.hpp>
